refactor(cli): move render helpers to render.go (PRD #1009 M1)
